Renamed InsertionsClient project to InsertionsClient.Console
