+++ src/InsertionsClient.Console/Api/DefaultConfigUpdater.cs
