+++ src/InsertionsClient.Console/Api/Props/Models/PropsFile.cs
