+++ src/InsertionsClient.Console/Api/Props/Models/PropsFileVariableReference.cs
