+++ src/InsertionsClient.Console/Api/Props/Models/PropsUpdateResults.cs
