+++ src/InsertionsClient.Console/Api/Props/Models/SwrFile.cs
