+++ src/InsertionsClient.Console/Api/Props/PropsFileUpdater.cs
