+++ src/InsertionsClient.Console/Api/Props/PropsVariableDeducer.cs
