+++ src/InsertionsClient.Console/Api/Props/SwrFileReader.cs
