+++ src/InsertionsClient.Console/Api/Providers/InsertionApiFactory.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System;
+
+namespace Microsoft.Net.Insertions.Api.Providers
+{
+    internal sealed class InsertionApiFactory : IInsertionApiFactory
+    {
+        public IInsertionApi Create(TimeSpan? maxWaitSeconds = null, TimeSpan? maxDownloadSeconds = null, int ? maxConcurrency = null)
+        {
+            return new InsertionApi(maxWaitSeconds, maxDownloadSeconds, maxConcurrency);
+        }
+    }
+}
