+++ src/InsertionsClient.Console/Common/Logging/Utilities.cs
