+++ src/InsertionsClient.Console/InsertionsClient.Console.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>Microsoft.Net.Insertions</RootNamespace>
+    <AssemblyName>InsertionsClient</AssemblyName>
+    <Nullable>enable</Nullable>
+    <WarningsAsErrors>nullable</WarningsAsErrors>
+  </PropertyGroup>
+
+
+  <ItemGroup>
+    <PackageReference Include="NuGet.Protocol" Version="5.5.1" />
+  </ItemGroup>
+
+</Project>
