+++ src/InsertionsClient.Console/Models/FileSaveResult.cs
