+++ src/InsertionsClient.Console/Models/PackageUpdateResult.cs
