+++ src/InsertionsClient.Console/Models/Update.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.Net.Insertions.Models
+{
+    internal enum Update
+    {
+        ExactMatch,
+        NoMatch,
+        Ignored,
+        NotAPackage
+    }
+}
