+++ src/InsertionsClient.Console/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "InsertionsClient": {
+      "commandName": "Project",
+      "commandLineArgs": "-d:C:\\temp\\insertions\\default.config -m:C:\\temp\\insertions\\manifest.json -i:C:\\temp\\insertions\\ignored.txt"
+    }
+  }
+}
