--- src/InsertionsClient/Models/Update.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-namespace Microsoft.Net.Insertions.Models
-{
-    internal enum Update
-    {
-        ExactMatch,
-        NoMatch,
-        Ignored,
-        NotAPackage
-    }
-}
